Add Docker support, client library, restructure packages
