+++ examples/mcp.json.windows
@@ -8,7 +8,7 @@
                 "--directory",
                 "C:\\ABSOLUTE\\PATH\\TO\\boomi-mcp-server",
                 "run",
-                "src/server.py"
+                "boomi_mcp_server/server.py"
             ]
         }
     }
