+++ server.py
@@ -1,4 +1,4 @@
-from src.server import main
+from boomi_mcp_server.server import main
 
 if __name__ == "__main__":
     main()
