+++ src/boomi_mcp_client/__init__.py
@@ -1,0 +1,4 @@
+"""Client library for interacting with Boomi MCP Server."""
+from .client import MCPClient
+
+__all__ = ["MCPClient"]
