+++ src/boomi_mcp_server/__main__.py
@@ -1,0 +1,4 @@
+from .server import main
+
+if __name__ == "__main__":
+    main()
